Add new features to upload the  profile photo of user and remove that photo.

+++ src/main/java/com/example/eCommercewebapp/model/dao/FileDataDAO.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface FileDataDAO extends JpaRepository<FileData,Integer> {
+public interface FileDataDAO extends JpaRepository<FileData,Long> {
     Optional<FileData> findByName(String fileName);
 }
